Fix UB in real to integer casts
(double)x may be greater or smaller than x; quoting the C standard
6.3.1.4:

"When a value of integer type is converted to a real floating type, if
the value being converted can be represented exactly in the new type,
it is unchanged. If the value being converted is in the range of
values that can be represented but cannot be represented exactly, the
result is either the nearest higher or nearest lower representable
value, chosen in an implementation-defined manner. If the value being
converted is outside the range of values that can be represented, the
behavior is undefined."

In the case of LLONG_MAX, compilers typically choose 2^63 because it
can be represented exactly. In this case, the range test in
json_cast_to_integer accepts the value 2^63, yielding UB in the double
to long long conversion. This commit fixes this issue and the
analogous issue in json_cast_to_unsigned by switching to the new fsdyn
functions to cast a number in IEEE 754 binary64 format to an integer.
This assumes that 'double' uses the binary64 format, which should hold
on all current architectures.